// backend/src/utils/types.ts
export enum CustomResponseStatus {
    SUCCESS = "success",
    FAIL = "fail"
}

export enum CustomResponseCodes {
    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
    TOKEN_VALID = "TOKEN_VALID",
    TOKEN_INVALID = "TOKEN_INVALID",
    TOKEN_EXPIRED = "TOKEN_EXPIRED",
    TOKEN_MALFORMED = "TOKEN_MALFORMED",
    TOKEN_NOT_PROVIDED = "TOKEN_NOT_PROVIDED",
    AUTHORIZED = "AUTHORIZED",
    UNAUTHORIZED = "UNAUTHORIZED",
    NOT_FOUND = "NOT_FOUND",
    UNPROCESSABLE_ENTITY = "UNPROCESSABLE_ENTITY"
}